Tarefa2_2019li1g104

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT2 :: [(Int, Jogada, Estado)]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

m1 :: [[Peca]]

Mapa teste

js1 :: [Jogador]

Lista teste de jogadores

Funções principais da Tarefa 2.

jogada

Arguments

:: Int
O identificador do Jogador que efetua a jogada.

-> Jogada
A Jogada a efetuar.

-> Estado
O Estado anterior.

-> Estado
O Estado resultante após o jogador efetuar a jogada.

Efetua uma jogada.

funcaoAcelera :: Int -> Estado -> Estado

Função para desecelerar

aumentaAcel :: Jogador -> Jogador

concretiza o efeito no estado da aceleração

funcaoDesacelera :: Int -> Estado -> Estado

Função para acelerar

diminuiAcel :: Jogador -> Jogador

concretiza o efeito no estado da desaceleração

Função para o movimento

funcaoMovimenta :: Int -> Direcao -> Estado -> Estado

Função de decisão do movimento a tomar

cima :: Int -> Jogador -> Estado -> Estado

Move para a pista à esquerda

baixo :: Int -> Jogador -> Estado -> Estado

Move para a pista á direita

esquerda :: Int -> Jogador -> Estado -> Estado

Diminui inclinação

direita :: Int -> Jogador -> Estado -> Estado

Aumenta inclinação

altEjGet :: Jogador -> Double

Devolve a altura de um estado ar

ecliEjGet :: Jogador -> Double

Devolve a inclinação de um estado ar

Função para disparo de cola

funcaoDispara :: Int -> Estado -> Estado

Concretiza o efeito do disparo no mapa e no estado do jogador

mudaPisoAnt :: Int -> Jogador -> Estado -> Estado

Altera o piso anterior e deconta uma munição

mudaPecAnt :: Jogador -> Estado -> Peca

Encontra uma peça no mapa e altera essa peca usando a função mudapeca

mudaPeca :: Peca -> Peca

Função que altera o piso de uma peça para cola

Funções para manipulação de listas e matrizes

type PosicaoMatriz = (Int, Int)

Defenição do tipo posição matriz

type Matriz a = [[a]]

Defenição do tipo matriz

encontraPosicaoMatriz :: PosicaoMatriz -> Matriz a -> a

Encontra o elemento numa dada posição de uma matriz

atualizaPosicaoMatriz :: PosicaoMatriz -> a -> Matriz a -> Matriz a

Dada uma posição de uma matriz atualiza essa posição com um novo elemento dado

atualizaIndiceLista :: Int -> a -> [a] -> [a]

Dado um indíce de uma lista atualiza essa esse índice com um novo elemento dado

encontraIndiceLista :: Int -> [a] -> a

Dado um indíce devolve o elemento correspondente

atualiza :: Int -> Jogador -> [Jogador] -> [Jogador]

Atualiza o jogador na lista de jogadores

Funções auxiliares

estaMorto :: Jogador -> Bool

Testa se o jogador vivo ou morto

estaNochao :: Jogador -> Bool

Testa se o jogador no chao ou nao

mostraJogador :: Int -> Estado -> Jogador

Recebe identificador devolve o jogador correspondente

temBalas :: Jogador -> Bool

Testa se o jogador tem balas ou não

Função de conversão

toGraus :: Double -> Double

Converte de radianos para graus

Funções de busca de alturas

getAltJ :: Peca -> Double -> Double -> Double

Devolve a altura exata no jogador na peça

getAltFPec :: Peca -> Double

Devolve a altura final de uma peça